Calibration in MM: Ticks looked up in FSM before dispense

+++ ui/dfuiserver.h
@@ -10,13 +10,13 @@ public:
     explicit DfUiServer(QObject *parent = 0);
     void startServer();
     void resetTimerSlot();
-    void updateVolumeSlot(int dispensed);
+    void updateVolumeSlot(double dispensed);
     void targetHitSlot();
     void initReadySlot();
 
 signals:
     void pleaseReset();
-    void updateVolume(int dispensed);
+    void updateVolume(double dispensed);
     void targetHit();
     void initReady();
